i18n/__init__.py
"""
Internationalization (i18n) system for the Physical AI & Humanoid Robotics textbook
Provides Urdu translation functionality as required by hackathon
"""
import os
from dotenv import load_dotenv

# Load environment variables
load_dotenv()

# Configuration for translation services
GOOGLE_TRANSLATE_API_KEY = os.getenv("GOOGLE_TRANSLATE_API_KEY")

def initialize_translation():
    """
    Initialize the translation system
    """
    print("Translation system initialized")
    # In a real implementation, this would connect to translation APIs
    pass

def translate_to_urdu(text):
    """
    Translate text to Urdu
    """
    # This would implement the actual translation logic
    # 1. Use Google Translate API if available
    # 2. Fallback to deep-translator if no API key
    # 3. Return translated text
    pass

def get_available_languages():
    """
    Return list of available translation languages
    Should include Urdu as required by hackathon
    """
    return ["ur", "en"]  # Urdu and English