requirements.txt
# ==========================================
# Physical AI & Humanoid Robotics Hackathon
# AI-Native Book + RAG + Demo Support
# ==========================================

# ---------- Backend API (FastAPI) ----------
fastapi>=0.104.1
uvicorn>=0.24.0
pydantic>=2.5.0
python-dotenv>=1.0.0
aiofiles>=23.2.1
python-multipart>=0.0.6

# ---------- Database (Neon Serverless Postgres) ----------
sqlalchemy>=2.0.23
asyncpg>=0.29.0

# ---------- Authentication (Better Auth compatible) ----------
cryptography>=41.0.0
passlib[bcrypt]>=1.7.4
python-jose>=3.3.0

# ---------- OpenAI Agents / RAG ----------
openai>=1.3.5
tiktoken>=0.5.2   # token counting for chunking & demos

# ---------- Vector Database ----------
qdrant-client>=1.7.0

# ---------- RAG Utilities ----------
numpy>=1.24.3

# ---------- Translation & Personalization (Demo-ready) ----------
deep-translator>=1.9.0

# ---------- Book & Content Processing ----------
markdown>=3.5.1
beautifulsoup4>=4.12.2
requests>=2.31.0

# ---------- Demo / Visualization (90-sec video friendly) ----------
Pillow>=10.0.0
matplotlib>=3.7.2
tqdm>=4.66.1

# ---------- Dev & Debug ----------
rich>=13.6.0

# Optional ML / heavy packages live in `requirements-ml.txt`

# npx create-docusaurus@latest physical-ai-book classic
# cd physical-ai-book
# npm install
